Remove emoji shapes
Unfortunately the implementation wasn't as far as I hoped. I'll have to revisit the implementation and find a good way to support emoji's and its different shapes and sizes in combination with rendering Text in a proper way. Right now I ran into some limitations properly rendering Text as a particle and rotate it for example.